requirements.txt
# Core dependencies
numpy>=1.20.0
pandas>=1.3.0
scikit-learn>=1.0.0
websocket-client>=1.2.0
PyQt6>=6.2.0
psutil>=5.9.0

# Testing dependencies
pytest>=6.2.5
pytest-asyncio>=0.16.0

# Development dependencies (optional)
black>=21.9b0
flake8>=3.9.0
mypy>=0.910
